Move ts refactor code to new branch
