--- frontend/src/metabase/components/Modal/index.tsx
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/no-default-export -- deprecated usage
-export { default } from "./Modal";
